rejex_pattern_simulator/pyproject.toml
[project]
name = "rejex-pattern-simulator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "automata-lib>=9.0.0",
    "graphviz>=0.20.3",
    "streamlit>=1.45.1",
]
